Move provenanced_addresses_from_filesystem_specs() back to engine/legacy/graph.py
It was improper for it to exist in build_files.py because this meant that build_files.py now needed to know about engine/legacy/graph.py, including Owners, OwnersRequest, and HydratedTargets. This is sort of like a circular dependency.
